Get methods for XML Parsing Classes, Addition of XML Test File
Introduced public get methods to allow access to component_data package
objects outside the package. Made minor fixes to aforementioned classes.
Added a new musicXML file to the test/resources/system package to allow
testing of XML parsing. Fixed minor typo for creating Notation objects
in the Note.java. Changed the test file for the Score.main() method
since we now have a file in the project, and the previous file was
located on my desktop.