[New Feature] Measure Number Displayed + Guitar Notes Selection
Displays the measure number on top of each note.
Guitar/Bass tabs only: Each note can be selected amd highlighted. The
last highlighted note is stored, and we can access its measure number
and its number among the notes in that measure to be used to implement
playing functionality from anywhere in the measure.